langchain-llm-apps/virtual-health-coach/requirements.txt
openai==1.5.0
langchain==0.0.342
python-dotenv==1.0.0
streamlit==1.28.2
packaging==23.2
faiss-cpu==1.7.4
youtube-transcript-api==0.6.1
nr-openai-observability

